requirements.txt
winput>=1.3.0
Pillow